+++ examples/todo-knex/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @keystonejs/demo-project-todo-knex
 
+## 5.1.23
+
+### Patch Changes
+
+- Updated dependencies [[`1200c3562`](https://github.com/keystonejs/keystone/commit/1200c356272ae8deea9da4267ce62c1449498e95), [`1200c3562`](https://github.com/keystonejs/keystone/commit/1200c356272ae8deea9da4267ce62c1449498e95)]:
+  - @keystonejs/keystone@18.0.0
+  - @keystonejs/app-graphql@6.2.0
+  - @keystonejs/adapter-knex@12.0.3
+
 ## 5.1.22
 
 ### Patch Changes