+++ packages/adapter-prisma/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @keystonejs/adapter-prisma
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`1200c3562`](https://github.com/keystonejs/keystone/commit/1200c356272ae8deea9da4267ce62c1449498e95), [`1200c3562`](https://github.com/keystonejs/keystone/commit/1200c356272ae8deea9da4267ce62c1449498e95)]:
+  - @keystonejs/keystone@18.0.0
+
 ## 1.1.0
 
 ### Minor Changes